Finish source refactoring
