+++ src/memory_protector/mac.cpp
