+++ src/memory_protector/win32.cpp
