+++ src/scanner/experimental/win32.cpp
